Compress and traversal function
Compression algorithm:
    Compress sequential nodes which do not have neither lokid nor hikid
but have key value to one single node. The key of the combined node
is a string with every key of sequential nodes. For example, nodes A, B,
C have 'a', 'b', 'c' key value. If all of them do not have any lokid or
hikid, then they can be shrinken into a single node which key value is
"abc" and has the same eqkid as node C.

    Traversing all the node by prefix search, and output the traversal
sequence of every word. The "raw_data.txt" is the word sequence
without compression, and the "compress_data.txt" is the word sequence
with compression.
    By using command "md5sum [filename]" on linux, we can get the
digests of the file. So that we can check the integrity of these files.